+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 
 ## 1.0.0
 - initial release
+
+## 1.0.1
+- Fixed fingerprints.
+- set_options is now a public method
+- changed option `write` to `output`